mining/gov/legislativo/esfera_estadual/alsp/requirements.txt
attrs==21.4.0
Automat==20.2.0
azure-core==1.24.2
azure-functions==1.11.2
azure-storage-blob==12.13.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
filelock==3.7.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isodate==0.6.1
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
msrest==0.7.1
oauthlib==3.2.0
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
-e git+https://github.com/joao-kitajima/poc-gov_datamining.git@3d7942f64aac1a985fd0bd26b5869c0b61bcd6aa#egg=themisminingcfg&subdirectory=mining/gov/legislativo/esfera_estadual/alsp/config
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.10
w3lib==1.22.0
zope.interface==5.4.0
